pyproject.toml
[project]
name = "hackai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
    "yt-dlp>=2025.6.9",
    "python-multipart>=0.0.6",
    "twelvelabs>=0.4.10",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
